… cpp/src/arrow/array-test.cc
… cpp/src/arrow/buffer-builder.h
… cpp/src/arrow/buffer-test.cc
… cpp/src/arrow/util/logging.h