This reduces code duplication.