@bkietz 2019-02-06
Hmmm, this is a performance regression.

Before:
```
BM_BuildPrimitiveArrayNoNulls/repeats:2_mean                 365772 us     365779 us          1     2.291GB/s
BM_BuildVectorNoNulls/repeats:2_mean                         555089 us     555097 us          1   922.362MB/s
BM_BuildBooleanArrayNoNulls/repeats:2_mean                   121138 us     121139 us          6   4.12748GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_mean                     13406 us      13406 us         52   4.66192GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_mean         28445 us      28445 us         25   2.19714GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_mean                    26330 us      26331 us         27   4.74726GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_mean        27448 us      27448 us         25   2.27696GB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean               136 us        136 us       5063   561.917MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean           125 us        125 us       5606   611.482MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean              126 us        126 us       5546   605.348MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                    317 us        317 us       2209   315.468MB/s
```

After:
```
BM_BuildPrimitiveArrayNoNulls/repeats:2_mean                 417623 us     417631 us          1   1.75068GB/s
BM_BuildVectorNoNulls/repeats:2_mean                         555594 us     555601 us          1   921.525MB/s
BM_BuildBooleanArrayNoNulls/repeats:2_mean                   730000 us     730006 us          1   701.364MB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_mean                     13394 us      13394 us         52   4.66611GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_mean         28451 us      28451 us         25    2.1967GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_mean                    26312 us      26313 us         27   4.75047GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_mean        27457 us      27457 us         25    2.2762GB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean               139 us        139 us       5040   549.399MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean           126 us        126 us       5570   606.613MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean              129 us        129 us       5406   589.165MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                    319 us        319 us       2193   313.216MB/s
```