@wesm 2019-02-14
BooleanBuilder seems to be more than 4 times slower with this patch

after

```
BM_BuildBooleanArrayNoNulls/repeats:2                        789455 us     789468 us          1   648.538MB/s
```

before

```
BM_BuildBooleanArrayNoNulls/repeats:2                        172299 us     172302 us          4   2.90189GB/s
```

Based on this evidence the iterator based unrolled bit generator vectorizes poorly. I just pushed changes https://github.com/apache/arrow/pull/3575/commits/be8442f8574cdc7d6df6f55d8f71136e217be894 reverting some of these changes (but still using `TypedBufferBuilder<bool>`, so better than before), as well as an option to disable tracking the false count, which we don't care about. For the record, disabling the false count only improved the benchmark from 650 MB/s to 1 GB/s, far less than the original 3GB/s. The compiler just isn't generating very good code for the CPU

If these changes are acceptable I suggest that we merge this patch (once the build is passing) and leave this be for now.