@wesm 2019-02-14
With my changes the perf seems slightly improved from the baseline

```
BM_BuildBooleanArrayNoNulls/repeats:2                        163077 us     162609 us          4   3.07486GB/s
```