@wesm 2019-02-14
I moved the entire `NumericBuilder<T>` implementation to the header file. I think this should make the linker problems go away. I don't expect an application to have a ton of instantiations of these templates so I don't know that it's worth stressing about code size for these until it can be demonstrated to be a problem. The size of `arrow-array-test` in a release build increased by about 3%.