@wesm 2019-02-15
I reverted some changes and think I fixed the valgrind / Windows / Python failures