@bkietz 2019-02-15
@wesm I'll try to write a unit test which produces the same error