@bkietz 2019-02-15
I can't reproduce https://travis-ci.org/apache/arrow/jobs/493915377#L2632 locally