@wesm 2019-02-16
Fixed the Ruby test failures and added a C++ unit test that failed. 

@xhochy I removed the `valid_bytes` argument from `AppendNulls`

https://github.com/apache/arrow/blob/master/cpp/src/arrow/array/builder_primitive.h#L55

If anyone was using this argument, it was not doing anything useful, unless the goal is to append zeroes. I'm not sure if this needs to go through a deprecation cycle. In our codebase it was only being used for internal purposes